requirements.txt
python-dotenv
browser-use
google-generativeai
pytrends
yfinance
pandas
streamlit
matplotlib
